requirements.txt
flask
python-dotenv
pydantic
google-genai
pillow
python-weather
psutil
schedule
pyautogui
SpeechRecognition
vosk
requests
beautifulsoup4
